# swagger-doc.json
{"openapi":"3.0.0","paths":{"/tags-svc/health":{"get":{"operationId":"HealthController_health","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["Health"]}},"/tags-svc":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/tags-svc/tag":{"post":{"operationId":"TagController_create","summary":"Post: Create a new tag.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagDto"}}}},"responses":{"201":{"description":"Tag has been successfully created. ","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagDto"}}}},"400":{"description":"Tag has not been created. Bad request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestException"}}}}},"tags":["Tag"]},"get":{"operationId":"TagController_getAll","summary":"Get: Get a list of all tags with conditions.","parameters":[{"name":"name","required":true,"in":"query","description":"name of tag","example":"revenue","schema":{"type":"string"}},{"name":"resource","required":false,"in":"query","description":"name of resource","example":"bank","schema":{"type":"string"}},{"name":"resourceId","required":false,"in":"query","description":"name of resourceId","example":"T001","schema":{"type":"string"}},{"name":"resourceType","required":false,"in":"query","description":"name of resourceType","example":"GST","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","description":"type of tag","example":"category","schema":{"type":"string"}},{"description":"name of tag","required":true,"example":"revenue","name":"name","in":"query","schema":{"type":"string"}},{"description":"name of resource","required":false,"example":"bank","name":"resource","in":"query","schema":{"type":"string"}},{"description":"name of resourceId","required":false,"example":"T001","name":"resourceId","in":"query","schema":{"type":"string"}},{"description":"name of resourceType","required":false,"example":"GST","name":"resourceType","in":"query","schema":{"type":"string"}},{"description":"type of tag","required":true,"example":"category","name":"type","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTagDto"}}}},"400":{"description":"Bad request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestException"}}}},"404":{"description":"No tag founded.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoTagFoundException"}}}}},"tags":["Tag"]}},"/tags-svc/tag/{id}":{"get":{"operationId":"TagController_getById","summary":"Get: Get a single tag by its ID.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Found tag successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseTagDto"}}}},"400":{"description":"Bad request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestException"}}}},"404":{"description":"No tag founded.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoTagFoundException"}}}}},"tags":["Tag"]},"patch":{"operationId":"TagController_update","summary":"Patch: Update existing tag with conditions","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTagDto"}}}},"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseTagDto"}}}},"204":{"description":"No tag founded.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoTagFoundException"}}}},"400":{"description":"Bad request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestException"}}}}},"tags":["Tag"]},"delete":{"operationId":"TagController_delete","summary":"Portal: Archive a single user by its ID.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatedByDto"}}}},"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"type":"string"}}}},"204":{"description":"No tag founded.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoTagFoundException"}}}},"400":{"description":"Bad request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestException"}}}}},"tags":["Tag"]}},"/tags-svc/tag/bulk-upload":{"post":{"operationId":"TagController_uploadFile","summary":"Post: Create tags from a csv file","parameters":[],"responses":{"201":{"description":"Successful","content":{"application/json":{"schema":{"type":"object"}}}},"400":{"description":"Bad request.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequestException"}}}}},"tags":["Tag"]}}},"info":{"title":"User Management Service API for Klubworks","description":"This swagger documentation includes the user service API for portal.","version":"1.0.0","contact":{}},"tags":[{"name":"Klub User Service","description":""}],"servers":[],"components":{"securitySchemes":{"Internal Auth0 API":{"type":"apiKey","description":"<b>Internal Auth0 Rule Key:</b>","name":"auth0-rule-key","in":"header"},"Internal API":{"type":"apiKey","description":"<b>Internal API Key:</b>","name":"internal-api-key","in":"header"}},"schemas":{"CreateTagDto":{"type":"object","properties":{"name":{"type":"string","description":"name of tag","example":"revenue"},"resource":{"type":"string","description":"name of resource","example":"bank"},"resourceId":{"type":"string","description":"name of resourceId","example":"T001"},"resourceType":{"type":"string","description":"name of resourceType","example":"GST"},"conditions":{"description":"tag conditions","example":[{"field":"description","keywords":["Amazon Vendor Central","Amazon Wholesale","Apple","Cash Deposit"],"condition":"contains"}],"type":"array","items":{"type":"array"}},"type":{"type":"string","description":"type of tag","example":"category"},"createdBy":{"type":"string","description":"identifier of tag creator","example":"6ebb0c45-f709-44a6-9fd9-476d1c9484e9"}},"required":["name","type","createdBy"]},"BadRequestException":{"type":"object","properties":{}},"ResponseTagDto":{"type":"object","properties":{"id":{"type":"string","description":"id of tag","example":"e8a5f497-584c-459e-87ca-2f0d5b770405"},"name":{"type":"string","description":"name of tag","example":"revenue"},"type":{"type":"string","description":"type of tag","example":"category"},"resource":{"type":"string","description":"name of resource","example":"bank"},"resourceId":{"type":"string","description":"name of resourceId","example":"T001"},"resourceType":{"type":"string","description":"name of resourceType of tag","example":"GST"},"conditions":{"description":"tag conditions","example":[{"field":"description","keywords":["Amazon Vendor Central","Amazon Wholesale","Apple","Cash Deposit"],"condition":"contains"}],"type":"array","items":{"type":"array"}},"isStatic":{"type":"boolean","description":"is tag static or not","example":true},"slug":{"type":"string","description":"slug of tag","example":"revenue-category-10Bank-GST-T000-4b109a5a-8eaa-4f75-91ae-1735212bb75b"}},"required":["id"]},"NoTagFoundException":{"type":"object","properties":{}},"GetTagDto":{"type":"object","properties":{"name":{"type":"string","description":"name of tag","example":"revenue"},"resource":{"type":"string","description":"name of resource","example":"bank"},"resourceId":{"type":"string","description":"name of resourceId","example":"T001"},"resourceType":{"type":"string","description":"name of resourceType","example":"GST"},"type":{"type":"string","description":"type of tag","example":"category"}},"required":["name","type"]},"UpdateTagDto":{"type":"object","properties":{"name":{"type":"string","description":"name of tag","example":"revenue"},"type":{"type":"string","description":"type of tag","example":"category"},"updatedBy":{"type":"string","description":"identifier of tag updator","example":"6ebb0c45-f709-44a6-9fd9-476d1c9484e9"},"resource":{"type":"string","description":"name of resource","example":"bank"},"resourceId":{"type":"string","description":"name of resourceId","example":"T001"},"resourceType":{"type":"string","description":"name of resourceType","example":"GST"},"conditions":{"description":"tag conditions","type":"array","items":{"type":"array"}}},"required":["name","type","updatedBy"]},"UpdatedByDto":{"type":"object","properties":{"updatedBy":{"type":"string","description":"identifier of tag updator","example":"6ebb0c45-f709-44a6-9fd9-476d1c9484e9"}},"required":["updatedBy"]}}}}